Refactor blocks module to submodules
