+++ modules/transport.js
@@ -374,7 +374,7 @@ Transport.prototype.internal = {
 		// According to maxium payload of 58150 bytes per block with every transaction being a vote
 		// Discounting maxium compression setting used in middleware
 		// Maximum transport payload = 2000000 bytes
-		modules.blocks.loadBlocksData({
+		modules.blocks.utils.loadBlocksData({
 			limit: 34, // 1977100 bytes
 			lastId: query.lastBlockId
 		}, function (err, data) {
